Create comparison_summary module
